Refactor: Table Filter Buttons 관련 상수 이동
