+++ fe/src/components/IssueTable/TableToolBar/IssueListFilters/TableFilterButtons/constant.js
@@ -1,0 +1,18 @@
+export const ASSIGNEE = {
+  KO: '담당자',
+  KEY: 'assignees',
+};
+export const LABELS = {
+  KO: '레이블',
+  KEY: 'labelNames',
+};
+export const MILESTONE = {
+  KO: '마일스톤',
+  KEY: 'milestoneName',
+};
+export const AUTHOR = {
+  KO: '작성자',
+  KEY: 'author',
+};
+
+export const ISSUE_TABLE_FILTER_OPTIONS = [ASSIGNEE, LABELS, MILESTONE, AUTHOR];
