+++ fe/src/constants.js
@@ -6,21 +6,3 @@ export const RESET = 'reset';
 export const ISSUE_STATE = 'issueState';
 
 export const NONE = '$none';
-export const ASSIGNEE = {
-  KO: '담당자',
-  KEY: 'assignees',
-};
-export const LABELS = {
-  KO: '레이블',
-  KEY: 'labelNames',
-};
-export const MILESTONE = {
-  KO: '마일스톤',
-  KEY: 'milestoneName',
-};
-export const AUTHOR = {
-  KO: '작성자',
-  KEY: 'author',
-};
-
-export const ISSUE_TABLE_FILTER_OPTIONS = [ASSIGNEE, LABELS, MILESTONE, AUTHOR];
